+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 7.02
  - Improved internet checking. Now ip address and host based methods
+ - Internet checking function splitted. Now separated from repository access
  
 ### 7.01
  - Fixed bug while changing interface mode if the destination interface name is in use